+++ libs/angular-components/tedi/components/layout/header/header.stories.ts
@@ -42,35 +42,35 @@ import { SeparatorComponent } from "../../helpers/separator/separator.component"
  */
 
 export default {
-  title: "TEDI-Ready Angular/Layout/Header",
+  title: "TEDI-Ready/Layout/Header",
   component: HeaderComponent,
   decorators: [
-      moduleMetadata({
-        imports: [
-          HeaderComponent,
-          HeaderContentComponent,
-          HeaderActionsComponent,
-          HeaderRoleComponent,
-          HeaderLanguageComponent,
-          HeaderProfileComponent,
-          HeaderLoginComponent,
-          HeaderLogoutComponent,
-          SeparatorComponent,
-          LinkComponent,
-          IconComponent,
-          ButtonComponent,
-          ShowAtDirective,
-          HideAtDirective,
-          SideNavComponent,
-          SideNavItemComponent,
-          SideNavToggleComponent,
-          SideNavOverlayComponent
-        ],
-      }),
-    ],
+    moduleMetadata({
+      imports: [
+        HeaderComponent,
+        HeaderContentComponent,
+        HeaderActionsComponent,
+        HeaderRoleComponent,
+        HeaderLanguageComponent,
+        HeaderProfileComponent,
+        HeaderLoginComponent,
+        HeaderLogoutComponent,
+        SeparatorComponent,
+        LinkComponent,
+        IconComponent,
+        ButtonComponent,
+        ShowAtDirective,
+        HideAtDirective,
+        SideNavComponent,
+        SideNavItemComponent,
+        SideNavToggleComponent,
+        SideNavOverlayComponent,
+      ],
+    }),
+  ],
   parameters: {
     layout: "fullscreen",
-    docs: { toc: false }
+    docs: { toc: false },
   },
   argTypes: {
     role: {
@@ -95,7 +95,7 @@ export default {
       table: {
         category: "header-role",
         type: { summary: "boolean" },
-        defaultValue: { summary: "false" }
+        defaultValue: { summary: "false" },
       },
     },
     representatives: {
@@ -130,19 +130,21 @@ export default {
       },
     },
     languages: {
-      description: "Languages object. <br />Key is value in 'Language' type. <br />Value should be text shown in the UI.",
+      description:
+        "Languages object. <br />Key is value in 'Language' type. <br />Value should be text shown in the UI.",
       table: {
         category: "header-language",
         type: { summary: "HeaderLanguage" },
       },
     },
-  }
+  },
 } as Meta<HeaderComponent>;
 
 export const Default: StoryObj<HeaderComponent> = {
   render: (args) => ({
     props: args,
-    styles: [`
+    styles: [
+      `
       img {
         max-height: 40px;
 
@@ -150,18 +152,19 @@ export const Default: StoryObj<HeaderComponent> = {
           max-height: 52px;
         }
       }
-    `],
+    `,
+    ],
     template: `
       <header tedi-header ${argsToTemplate(args)}>
         <button tedi-sidenav-toggle></button>
         <img src="header-logo.svg" alt="Logo" />
         <tedi-header-content *showAt="'lg'">
-          <a href="#">Link text</a>
-          <a href="#">Link text</a>
+          <a tedi-link href="#" [underline]="false">Link text</a>
+          <a tedi-link href="#" [underline]="false">Link text</a>
         </tedi-header-content>
         <tedi-header-actions>
-          <tedi-header-language 
-            [languages]="{ et: 'EST', en: 'ENG', ru: 'RUS' }" 
+          <tedi-header-language
+            [languages]="{ et: 'EST', en: 'ENG', ru: 'RUS' }"
             currentLanguage="et"
           />
           <tedi-header-login />
@@ -183,7 +186,8 @@ export const Default: StoryObj<HeaderComponent> = {
 export const LoggedIn1: StoryObj<HeaderComponent> = {
   render: (args) => ({
     props: args,
-    styles: [`
+    styles: [
+      `
       img {
         max-height: 40px;
 
@@ -191,15 +195,16 @@ export const LoggedIn1: StoryObj<HeaderComponent> = {
           max-height: 52px;
         }
       }
-    `],
+    `,
+    ],
     template: `
       <div style="height: 300px;">
         <header tedi-header ${argsToTemplate(args)}>
           <img src="header-logo.svg" alt="Logo" />
           <tedi-header-actions>
-            <tedi-header-role 
+            <tedi-header-role
               role="Roll:"
-              description="49504080934" 
+              description="49504080934"
               [showInput]="true"
               [representatives]="[
                 { icon: 'person', name: 'Juulia Sarapuu', description: '62004122984' },
@@ -211,9 +216,9 @@ export const LoggedIn1: StoryObj<HeaderComponent> = {
             />
             <tedi-header-language [languages]="{ et: 'EST', en: 'ENG', ru: 'RUS' }" />
             <tedi-header-profile showDropdown="lg">
-              <tedi-header-role 
+              <tedi-header-role
                 role="Roll:"
-                description="49504080934" 
+                description="49504080934"
                 [showInput]="true"
                 [representatives]="[
                   { icon: 'person', name: 'Juulia Sarapuu', description: '62004122984' },
@@ -241,7 +246,8 @@ export const LoggedIn1: StoryObj<HeaderComponent> = {
 export const LoggedIn2: StoryObj<HeaderComponent> = {
   render: (args) => ({
     props: args,
-    styles: [`
+    styles: [
+      `
       img {
         max-height: 40px;
 
@@ -249,15 +255,16 @@ export const LoggedIn2: StoryObj<HeaderComponent> = {
           max-height: 52px;
         }
       }
-    `],
+    `,
+    ],
     template: `
       <div style="height: 300px;">
         <header tedi-header ${argsToTemplate(args)}>
           <img src="header-logo.svg" alt="Logo" />
           <tedi-header-actions>
-            <tedi-header-role 
+            <tedi-header-role
               role="Roll:"
-              description="49504080934" 
+              description="49504080934"
               [showInput]="true"
               [representatives]="[
                 { icon: 'person', name: 'Juulia Sarapuu', description: '62004122984' },
@@ -281,7 +288,8 @@ export const LoggedIn2: StoryObj<HeaderComponent> = {
 export const LoggedIn3: StoryObj<HeaderComponent> = {
   render: (args) => ({
     props: args,
-    styles: [`
+    styles: [
+      `
       img {
         max-height: 40px;
 
@@ -289,15 +297,16 @@ export const LoggedIn3: StoryObj<HeaderComponent> = {
           max-height: 52px;
         }
       }
-    `],
+    `,
+    ],
     template: `
       <div style="height: 300px;">
         <header tedi-header ${argsToTemplate(args)}>
           <img src="header-logo.svg" alt="Logo" />
           <tedi-header-actions>
-            <tedi-header-role 
+            <tedi-header-role
               role="Roll:"
-              description="49504080934" 
+              description="49504080934"
               [showInput]="true"
               [representatives]="[
                 { icon: 'person', name: 'Juulia Sarapuu', description: '62004122984' },
@@ -309,9 +318,9 @@ export const LoggedIn3: StoryObj<HeaderComponent> = {
             />
             <tedi-header-language [languages]="{ et: 'EST', en: 'ENG', ru: 'RUS' }" />
             <tedi-header-profile name="Juulia Sarapuu" showDropdown="lg">
-              <tedi-header-role 
+              <tedi-header-role
                 role="Roll:"
-                description="49504080934" 
+                description="49504080934"
                 [showInput]="true"
                 [representatives]="[
                   { icon: 'person', name: 'Juulia Sarapuu', description: '62004122984' },
@@ -339,7 +348,8 @@ export const LoggedIn3: StoryObj<HeaderComponent> = {
 export const LoggedOut: StoryObj<HeaderComponent> = {
   render: (args) => ({
     props: args,
-    styles: [`
+    styles: [
+      `
       img {
         max-height: 40px;
 
@@ -347,7 +357,8 @@ export const LoggedOut: StoryObj<HeaderComponent> = {
           max-height: 52px;
         }
       }
-    `],
+    `,
+    ],
     template: `
       <header tedi-header ${argsToTemplate(args)}>
         <img src="header-logo.svg" alt="Logo" />
@@ -372,8 +383,8 @@ export const WithOrganization: StoryObj<HeaderComponent> = {
               Ligipääsetavus
               <tedi-icon name="call_made" />
             </a>
-            <tedi-header-role 
-              role="Asutus:" 
+            <tedi-header-role
+              role="Asutus:"
               [showInput]="true"
               [representatives]="[
                 { icon: 'person', name: 'Valik 1' },
@@ -383,9 +394,9 @@ export const WithOrganization: StoryObj<HeaderComponent> = {
               [currentRepresentative]="{ icon: 'person', name: 'Valik 1' }"
               *showAt="'lg'"
             />
-            <tedi-header-role 
+            <tedi-header-role
               role="Roll:"
-              description="49504080934" 
+              description="49504080934"
               [showInput]="true"
               [representatives]="[
                 { icon: 'person', name: 'Juulia Sarapuu', description: '62004122984' },
@@ -397,8 +408,8 @@ export const WithOrganization: StoryObj<HeaderComponent> = {
             />
             <tedi-header-language [languages]="{ et: 'EST', en: 'ENG', ru: 'RUS' }" />
             <tedi-header-profile showDropdown="lg">
-              <tedi-header-role 
-                role="Asutus:" 
+              <tedi-header-role
+                role="Asutus:"
                 [showInput]="true"
                 [representatives]="[
                   { icon: 'person', name: 'Valik 1' },
@@ -408,9 +419,9 @@ export const WithOrganization: StoryObj<HeaderComponent> = {
                 [currentRepresentative]="{ icon: 'person', name: 'Valik 1' }"
                 *hideAt="'lg'"
               />
-              <tedi-header-role 
+              <tedi-header-role
                 role="Roll:"
-                description="49504080934" 
+                description="49504080934"
                 [showInput]="true"
                 [representatives]="[
                   { icon: 'person', name: 'Juulia Sarapuu', description: '62004122984' },