+++ libs/angular-components/tedi/components/navigation/link/link.stories.ts
@@ -6,8 +6,8 @@ import {
   StoryFn,
 } from "@storybook/angular";
 import { LinkComponent } from "./link.component";
-import { ColComponent } from "../../layout/grid/col/col.component";
-import { RowComponent } from "../../layout/grid/row/row.component";
+import { ColComponent } from "../../helpers/grid/col/col.component";
+import { RowComponent } from "../../helpers/grid/row/row.component";
 import { TextComponent, TextColor } from "../../base/text/text.component";
 import { IconComponent } from "../../base/icon/icon.component";
 import { createBreakpointArgTypes } from "../../../../dev-tools/createBreakpointArgTypes";
@@ -20,7 +20,7 @@ const PSEUDO_STATE = ["Default", "Hover", "Active", "Focus"];
  */
 
 export default {
-  title: "TEDI-Ready Angular/Navigation/Link",
+  title: "TEDI-Ready/Components/Navigation/Link",
   component: LinkComponent,
   decorators: [
     moduleMetadata({